子クラスis親クラスになるようにクラス分けを修正、クラス名を修正
